Making pie graphs more useful.
Previously targets had to be in the form of target=name:value
Now targets can be targets simliar to line graphs. The default behavior is to find the average value of the series and use that as the size of the slice. 'from' and 'until' can be used to adjust the averaging window.
The mode can be changed with the pieMode variable. Current options are average, maximum, and minimum but can be extended by adding functions to functions.py