Always compact hashes to create keys for memcached
Currently, the code compacts if the length of the request key is > 249.
The memcache max key length is 250, so this would appear to make sense.
Unfortunately, under the covers, django's memcached handling creates a
key by combining it "with the cache prefix and key version to provide
a final cache key".

See for more:
https://docs.djangoproject.com/en/dev/topics/cache/#cache-key-transformation