fix holt winters bug of mismatched units
also make holt winters calculations much simpler and generalize
bootstrap code such that other functions could use it

when fetching the data for the bootstrap, it could get bootstrap data
with different units of time, which caused index errors and stack traces.
now it fetches the bootstrap and original data in one series so they're
all using the same units of time.