fix log files ending up in LOG_DIR's parent
If log-dir is not set on the commandline, the logdir used would be "%s-%s" % (LOGDIR, instance)
This meant that the actual log dir would be in the same parent of
LOG_DIR but would be named differently than specified. It also caused a
conflict when multiple daemons shared a common LOG_DIR - e.g.
carbon-cache and carbon-aggregator instance 'a' would both end up in
/var/log/graphite-a/ for a LOG_DIR setting of /var/log/graphite and would have
conflicting console.log files
Now, in that case the logs will end up in /var/log/graphite/carbon-cache-a/ and
/var/log/graphite/carbon-aggregator-a/