Update the documentation for derivative()
Two issues: one is that it doesn't necessarily calculate per-minute
rates (since that depends on the aggregation function). The other is
that a true derivative is a rate of change over time, as opposed to
just rate of change between subsequent data points. Add in a reference
to the perSecond function, since changing the definition of 'derivative'
now might be considered a breaking change.