Make sure default options are properly used in glyph.py
This fixes #314. This makes it so that the defaultGraphOptions are
always set as the default when a graphTemplates.conf exists and any
fields that the specific config overrides will trump the defaults.

It also updates the default profile in graphTemplates.conf to not
be missing keys. The lineColors were taken directly from glyph.py

This also prevents a LOT of this type of mess a few lines down:
    defaults.get('foo', defaultGraphOptions['foo'))