Enable serving graphite.dashboard app outside root
Previously, serving dashboard app from anywhere but
/dashboard/ would break things because there
are hardcoded absolute paths in dashboard.js.

This change adds settings.GRAPHITE_API_PREFIX, to
allow use of the dashboard app at some other place
in the URL namespace.

This is useful if we want to deploy graphite along with
other WSGI applications, and use a /graphite/ prefix
to namespace it away from other apps.