Fix 'now' handling in render queries
Previously, datetime.now() was called more than once
when rendering data.  This led to non-deterministic
selection of sample resolution.  For example, if
the schema is 60s:1d,300s:7d, and a request for
'-1d'->'now' came in, one would sometimes get
the 300s data and sometimes the 60s data, depending
on the length of time between the call to now() in
render.views and the call to now() in whisper.fetch.