Make divideSeries accept multiple divisor series
Remove the current limitation of divisorSeriesList to have only one
series and accept multiple series. In this case the number of divisor
series must match the number of dividend series, and they are matched so
each dividend series get divided by the corresponding divisor.

The old behaviour when only one series is used is retained, so this is
only an extension.