Add mapSeries and reduceSeries render functions
Currently, graphite has no way to map-reduce disparate series together
which is desirable for series that come from collectd and others
where calculations need to be performed on two or more related series.

This commit adds two new render functions: mapSeries and reduceSeries.
This is a generic way of mapping several related series together and
reducing them with another function.

- Add mapSeries function which takes a seriesList and maps it to a
  list of sub-seriesList. Each sub-seriesList has the given mapNode
  in common.
- Add reduceSeries function which takes a list of seriesLists and
  reduces it to a list of series by means of the reduceFunction.
- Add mock module to .travis.yml
- Add tests for mapSeries and reduceSeries