Clarify Python/Django requirements
Nothing really prevents Django 1.3 from working but the tests won't run with
it -- it doesn't have `override_settings()`.

Some code already prevents Graphite-web from running on Python 2.5. Travis
doesn't even support it so it's hopeless to claim 2.5 support.

This patch syncs requirements with what we actually test on Travis, which is
the sanest approach.