Set the step of the timeFunction() to 10 seconds.
As far as I could tell, there was no reason that this was set at 60 seconds instead of 10 seconds.  10 seconds is the default resolution for most stats stored in Carbon, so it makes sense to match that.

If someone needs to go back to 60 second intervals, `alias(summarize(time(), "60sec", "min"), "The.time.series")` would do the trick.