Use OrderedDict in groupByNode
Checking `key` for membership in `metaSeries.keys()` is inefficient and
unnecessary since we can use the `setdefault` method to replace that
logic.

Additionally, the `keys` list appears to only be necessary for the
purposes of maintaining ordering; this can be managed instead by using
an OrderedDict instead of a dict and a list.