Backport of #1188 "Serve staticfiles in the webapp if whitenoise is installed" to 0.9.x
This avoids the collectstatic step & apache configuration. It's probably slightly less optimal than serving files with apache/nginx but much easier to start with.
Whitenoise is compatible with Django 1.4 and above, but before version 2.0.2 it was compatible only with python 2.7, so in this version safecheck is included.

Tested with synthesize and manual whitenoise installation, but please check more.