add optional parameter referenceSeries to transformNull
This commit adds an optional third parameter, referenceSeries, to
transformNull().  This argument, if specified, governs that
transformation of nulls should not happen unconditionally on every
time interval, but rather, on the condition that the same time
interval in any series of referenceSeries is also non-null.

The motivation for this functionality stems from the need to
differentiate between nulls that represent "zero" and those that
represent "no data".  The parameter referenceSeries is used to
identify the time intervals to which there are data, so that
transformNulls() can transforms nulls in such intervals as
unequivocal zeros.