replace _fetchWithBootstrap
I discovered that movingAverage() on data that was already retrieved by
a non-summary function like perSecond() was going back to retrieve
earlier data, which it needs to do to get the first points in the graph
correct. However in retrieving this data it was not using the original
wildcard but polling each datapoint individually. On queries with over a
thousand datapoints in a clustered environment this was incredibly
resource intensive and slow.

This change passes the original tokens to each function as an element
'args' in the resourceContext. So if a function needs to retrieve the
original data with a time offset it can do so, and retrieve the data in
the same way as it was originally retrieved with wildcards intact. This
is effecient.

Four functions used _fetchWithBootstrap, so I updated all of them to use
the new mechanism. There are a bunch of other function that also re-pull
data but do not use _fetchWithBootstrap. I believe they could also
benefit from this, but I have not addressed them.

I verified the output against the original function results and did find
some discrepancies, but I believe these are bugs in the old code, so the
results should now be better as well as faster. These discrepancies
were:

In movingMedian _fetchWithBootsstap sometimes inserted an incorrect Null
before the time window, changing the result for the first few points

holtWinters functions old version had a discontinuity around a 7 day
time selection. 167 hours significantly different from 169 hours. This
discontinuity is now gone, but the results for shorter time windows are
now significantly changed. I'm not a statistician but I suspect these
are intended for long time windows and the discontinuity was a bug
brought on by mixing two whisper ranges together.

Conflicts:
	webapp/graphite/render/functions.py