Handle missing remote stores more gracefully
When finding remote metrics, a single unavailable backend causes the
entire tree to be unavailable for two consecutive requests, as
graphite/metrics/views.py's find_view() raises an AttributeError rather
than marking the backend as failed, and continuing with others.

This is ultimately because remote_storage.py's get_results() fails to
check for errors before trying to work with the response from httplib.

This patch enables all other backends' metrics to be shown immediately
after a backend fails.