Attempt at making histograms useful in Graphite.
Currently, even if one exports an histogram to Graphite, it is
still relatively difficult to use as there is no way to easilly display
it or get humanly understandable data out of it.

With summaryForHistgram() you'll get the usual min, max, sum, count
and a few percentiles to try to make sense of your dashboard.

Histograms are expected to be a list of series which contains their
bucket upper_bound in a node, for example:

       latency.histogram.10 - values < 10ms
       latency.histogram.20 - values < 20ms
       latency.histogram.inf - values < inf

Histogram can be cummulative or not. If they are the value is the
sum of everything under the upper bound, else this stops at the previous
upper bound.

Histograms can be exported as proper monotonic counters (to be then
fed to summaryForHistogram() after applying nonNegativeDerivative())
but simple gauges can also be used (not recommended!).