Forward HTTP request headers to CLUSTER_HOSTS
When making requests to remote webapps (defined in CLUSTER_HOSTS), allow us to
forward a configurable set of HTTP headers along from the original request.

The REMOTE_STORE_FORWARD_HEADERS setting, defaulting to an empty list, takes an
iterable of header names that are forwarded from the originating request along
to remote webapps, when a remote store request is made.

In our setup, this allows us to forward cookies and so preserve SSO auth status
between all graphite servers in a render cluster

Port of #1733 to master