Manipulate graph targets and implement New Graph -> New Empty Graph (#424)
Buttons to re-order and remove graph targets on the Graph Target grid.

Button to add a new target to a graph.

Up and Down buttons are from http://tango.freedesktop.org/.  Delete button already existed in the source.

Force the GridPanel to fit.  Makes the images look better.

Add in New Empty Graph.

Remove calls to addTargetToSelectedGraph as the listeners on the targetStore invoke them.

Fix name of delete icon to be the new trash icon.

Adjust location of icons to match new locations.