Fix hash position computation of fnv1a_ch.
If the fnv1a hash of a given key returns a value less than 0x10000000,
then the high and low parts of the hash would be split incorrectly,
resulting in a wrongly computed hash position.

This removes the hex formatting and splitting of the fnv1a hash, and
just calculates the position using bit manipulation.

Patch taken from graphite-project/carbon#715.