Update config-carbon.rst
I tried to change the config file 'storage-aggregation.conf', add a section at the bottom, and it takes no effect. So after I read all the documents, I think the order of sections matters. So  I think it is important to add the order explanation to the 'storage-aggregation.conf' document sections. 

I have trapped by this order problem for a while.