Default GRAPHITE_ROOT to sys.prefix
The GRAPHITE_ROOT was to be set to the parent directory of the path where
graphite-web is installed. The `storage`, `static` and `content` directories are
set to be dubdirs of GRAPHITE_ROOT.

If installing Graphite in the regular Python path's this would make the directories
look something like `/usr/lib/python3.7/{storage,static,conf}`. Which is probably not
what the user wants.

This sets the default to `sys.prefix`.

sys.prefix