fix function parameter types
sets parameter type of functions which represent a data point value from
"integer" to "float" if they were "integer" before. in cases where the
documentation says that the type is "integer" it updates it to "constant"
("float" or "integer").

in cases where a parameter represents a percentile this also sets it to
"float", previously most of them were set to "integer".

also changes first parameter of "useSeries()" from "string" to "float",
because I think it should be.