better error messages
with this change the error message includes useful hints regarding what
the type of a given invalid value is.
F.e. previously a message talking about an int(1) value would have been
the same as a message talking about a str('1'), but now they are
differentiable because repr() indicates the type