flake8: re-enable some rules which need the least churn to fix
see removed lines from setup.cfg for the specific rules re-enabled

The N8xx ignored rules were for the "pep8-naming" plugin for flake8.
It is not installed by default, it probably has other warnings
that this codebase would trigger, and there are probably other
flake8 plugins that would cause lots of warnings, so just remove.