Fix bool() issue in recursion termination check.
The behaviour of ParsedResults seems to have changed in the way that
the usage of pop() along with the bool() check on the object interact
wrt checking if there are further elements in the parsed structure to
consume. In version > 3.0 bool() checks whether either the internal
_toklist or _tokdict members contain items left to consume and returns
True if either do, whereas pop with no arguments will only consume from
the underlying list. That means that we would get a True in the if
condition in this code, whereas a subsequent call to pop() would throw
an exception. Calling asList() forces the use of the list representation
in the bool check here and so fixes this issue.